requirements.txt.txt
faiss-cpu
sentence-transformers
transformers
torch
streamlit
numpy
pandas